// src/types/view-decline-cases-type.ts
export interface declineLeadCases {
  leadCaptureId: number;
  name: string;
  email: string;
  phone: string;
  coreAcademicCareerId: number;
  careerName: string;
  coreAcademicProgramId: number;
  programName: string;
  coreLeadStageId: number;
  leadStageDescription: string;
  coreLeadSubStageId: number;
  leadSubStageDescription: string;
  modifiedOn: string | any;
  leadSourceName: string;
  status: string | any;
  createdAt: string | any;
  leadOwner: string | any;
}